+++ src/Domain/Metric/ClassBased/WMC.php
@@ -6,9 +6,12 @@ namespace PHPQuality\Domain\Metric\ClassBased;
 
 use PHPQuality\Domain\Metric;
 
+/**
+ * @psalm-external-mutation-free
+ */
 final class WMC extends ClassBasedMetric
 {
-    /** @var int */
+    /** @var float */
     private $wmc = 0;
 
     /**
@@ -20,7 +23,7 @@ final class WMC extends ClassBasedMetric
         $this->wmc += $metric->calculate();
     }
 
-    public function calculate(): int
+    public function calculate(): float
     {
         return $this->wmc;
     }
