+++ src/Domain/Metric/MethodBased/CC.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace PHPQuality\Domain\Metric\MethodBased;
 
+/**
+ * @psalm-external-mutation-free
+ */
 final class CC extends MethodBasedMetric
 {
     /** @var int */
@@ -17,7 +20,7 @@ final class CC extends MethodBasedMetric
         $this->complexity++;
     }
 
-    public function calculate(): int
+    public function calculate(): float
     {
         return $this->complexity;
     }
